add Graph class in sql.dart
