+++ lib/light.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:handrange/combination.dart';
+import 'package:handrange/sql.dart';
 
 class Light extends ChangeNotifier {
   List<Map<String, dynamic>> status = CONBI.map((e) => {
@@ -20,14 +21,11 @@ class Light extends ChangeNotifier {
   double count = 0;
 
   onInsert() async {
-    Hand hand = Hand(status: status);
-    await Hand.insertMemo(hand);
+    Graph graph = Graph(status: status);
+    await Graph.insertGraph(graph);
   }
 
-  onGet() async {
-    final List<Hand> hands = await Hand.getMemos();
-    print(await hands());
-  }
+
 
   onTapped(String hand) {
     status.forEach((element) {
@@ -111,24 +109,3 @@ class Light extends ChangeNotifier {
 
 }
 
-class Dog {
-  final List<Map<String, dynamic>> status;
-
-  Dog({this.status});
-  
-  Map<String, dynamic> toMap() {
-    Map<String, bool> s;
-    status.forEach((element) {
-      s[element["hand"]] = element["isSelected"]; // s["AA"] = false;
-    });
-
-    return s;
-  }
-
-  // Implement toString to make it easier to see information about
-  // each dog when using the print statement.
-  @override
-  String toString() {
-    return 'Dog{status: $status}';
-  }
-}
